Move triangle class logic to its own variable
